Moved the listener from index.js to events/interactionCreate.js
